feat: 투표 기능 구현

+++ data/src/main/java/com/susu/data/remote/model/response/PopularVoteResponse.kt
@@ -10,7 +10,7 @@ data class PopularVoteResponse(
     val id: Long,
     val boardName: String,
     val content: String,
-    val count: Int,
+    val count: Long,
     val isModified: Boolean,
 )
